Refactor Move openModal and closeModal functions to Modal.js. Use Refs in PortfolioSection.js
