+++ src/components/PortfolioSection.js
@@ -2,38 +2,15 @@ import React from "react"
 import styled from "styled-components";
 import Profile from "./ProfilePic"
 import Modal from "./Modal"
-import gsap from "gsap"
 
 class Portfolio extends React.Component {
   constructor(props) {
     super(props)
-    this.state = { modalOpen: false }
-  }
-
-  openModal() {
-    document.body.style.height = "100vh"
-    document.body.style.overflow = "hidden"
-    var tl = gsap.timeline()
-    tl.to(`#${this.props.name}_modal`.replace(/\s/g, ""), {duration: 0.15, autoAlpha: 1})
-    tl.to(`#${this.props.name}_modalContentWrapper`.replace(/\s/g, ""), {duration: 0.25, opacity: 1, scale: 1, ease: "back"})
-    this.setState({ modalOpen: true })
-  } 
-
-  closeModal() {
-    var tl = gsap.timeline()
-    tl.to(`#${this.props.name}_modalContentWrapper`.replace(/\s/g, ""), {duration: 0.25, opacity: 0, scale: 0, ease: "power2"})
-    tl.to(`#${this.props.name}_modal`.replace(/\s/g, ""), {duration: 0.15, autoAlpha: 0})
-    document.body.style.height = null
-    document.body.style.overflow = null
-    this.setState({ modalOpen: false })
+    this.modalRef = React.createRef()
   }
 
   handleClick() {
-    if (!this.state.modalOpen) {
-      this.openModal()
-    } else {
-      this.closeModal()
-    }
+    this.modalRef.current.openModal()
   }
 
   render() {
@@ -46,6 +23,7 @@ class Portfolio extends React.Component {
           name={this.props.name} 
           content={this.props.children} 
           marble={this.props.image}
+          ref={this.modalRef}
         />
       </PortfolioWrapper>
     )
